Converted other routes to new design.
