+++ src/app/(typef)/games/[game]/servers/[serverId]/commands/start-server/start-server.tsx
