+++ src/app/(typef)/games/[game]/servers/[serverId]/commands/stop-server/stop-server.action.ts
