+++ src/app/(typef)/games/[game]/servers/[serverId]/layout.tsx
@@ -1,7 +1,33 @@
-export default function GamesLayout({
+import ServerSideBar from "@/(global)/components/server/server-side-bar";
+
+const routes = [
+    {
+        label: "Overview",
+        path: "overview",
+    },
+    {
+        label: "Commands",
+        path: "commands",
+    },
+    // {
+    //     label: "Settings",
+    //     path: "settings",
+    // },
+    // {
+    //     label: "Shop",
+    //     path: "shop",
+    // },
+];
+
+export default function Layout({
     children,
 }: Readonly<{
     children: React.ReactNode;
 }>) {
-    return <>{children}</>;
+    return (
+        <ServerSideBar routes={routes} serverName="Template">
+            {" "}
+            {children}{" "}
+        </ServerSideBar>
+    );
 }
